Reorganise into shallow directory structure
