+++ features/step_definitions/a11y_steps.js
@@ -2,7 +2,7 @@ const { defineSupportCode } = require('cucumber')
 var assert = require('assert')
 var mkdirp = require('mkdirp')
 var rimraf = require('rimraf')
-var Standards = require('../../lib/bbc/a11y/js/standards')
+var Standards = require('../../lib/standards')
 var jquery = require('jquery')
 var webServer = require('../support/web_server')
 var childProcess = require('child_process')
