+++ lib/reporter.js
