+++ lib/standards/imageAlternatives/imagesMustHaveAltAttributes.js
