+++ lib/standards/index.js
@@ -1,4 +1,4 @@
-var xpath = require('./xpath');
+var xpath = require('../xpath');
 
 function Standards(standards, skipped, only) {
   this.standards = standards;
@@ -9,45 +9,45 @@ function Standards(standards, skipped, only) {
 Standards.sections = {
 
   focusableControls: [
-    require('./standards/focusableControls/anchorsMustHaveHrefs')
+    require('./focusableControls/anchorsMustHaveHrefs')
   ],
 
   formInteractions: [
-    require('./standards/formInteractions/formsMustHaveSubmitButtons')
+    require('./formInteractions/formsMustHaveSubmitButtons')
   ],
 
   formLabels: [
-    require('./standards/formLabels/fieldsMustHaveLabelsOrTitles')
+    require('./formLabels/fieldsMustHaveLabelsOrTitles')
   ],
 
   headings: [
-    require('./standards/headings/contentMustFollowHeadings'),
-    require('./standards/headings/exactlyOneMainHeading'),
-    require('./standards/headings/headingsMustBeInAscendingOrder')
+    require('./headings/contentMustFollowHeadings'),
+    require('./headings/exactlyOneMainHeading'),
+    require('./headings/headingsMustBeInAscendingOrder')
   ],
 
   imageAlternatives: [
-    require('./standards/imageAlternatives/imagesMustHaveAltAttributes'),
+    require('./imageAlternatives/imagesMustHaveAltAttributes'),
   ],
 
   indicatingLanguage: [
-    require('./standards/indicatingLanguage/htmlMustHaveLangAttribute')
+    require('./indicatingLanguage/htmlMustHaveLangAttribute')
   ],
 
   mainLandmark: [
-    require('./standards/mainLandmark/exactlyOneMainLandmark')
+    require('./mainLandmark/exactlyOneMainLandmark')
   ],
 
   minimumTextSize: [
-    require('./standards/minimumTextSize/textCannotBeTooSmall')
+    require('./minimumTextSize/textCannotBeTooSmall')
   ],
 
   tabIndex: [
-    require('./standards/tabIndex/elementsWithZeroTabIndexMustBeFields')
+    require('./tabIndex/elementsWithZeroTabIndexMustBeFields')
   ],
 
   titleAttributes: [
-    require('./standards/titleAttributes/titleAttributesOnlyOnInputs')
+    require('./titleAttributes/titleAttributesOnlyOnInputs')
   ]
 
 }
