+++ lib/standards/indicatingLanguage/htmlMustHaveLangAttribute.js
