+++ lib/standards/mainLandmark/exactlyOneMainLandmark.js
