+++ lib/xpath.js
