+++ package.json
@@ -6,8 +6,8 @@
   "directories": {},
   "scripts": {
     "test": "./node_modules/.bin/electron-mocha --renderer && ./node_modules/.bin/cucumber-electron",
-    "browserify": "browserify ./lib/bbc/a11y/js/a11y.js -o ./lib/bbc/a11y/js/bundle.js",
-    "watchify": "watchify ./lib/bbc/a11y/js/a11y.js -o ./lib/bbc/a11y/js/bundle.js"
+    "browserify": "browserify ./lib/a11y.js -o ./dist/bundle.js",
+    "watchify": "watchify ./lib/a11y.js -o ./dist/bundle.js"
   },
   "bin": {
     "bbc-a11y": "./bin/bbc-a11y.js"
